+++ src/components/TokenList/index.ts
@@ -1,0 +1,2 @@
+export { TokenList } from "./TokenList";
+export type { TokenListItem, TokenListProps } from "./TokenList";